Moved Influx DB calls to new controller and deleted the old one
